+++ sound-mind-studio/tests/test_landing_page.h
@@ -6,6 +6,7 @@ class LandingPageTest : public QObject {
     Q_OBJECT
 
 private slots:
+    void showsTheEmbeddedLogo();
     void newProjectButtonEmitsNewProjectRequested();
     void openProjectButtonEmitsOpenProjectRequested();
     void setRecentProjectsWithNoPathsShowsNoClickableEntries();