+++ sound-mind-studio/tests/test_theme.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <QObject>
+
+class ThemeTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void studioStyleSheetIsNotEmpty();
+    void studioStyleSheetContainsBothBrandColors();
+    void studioWindowIconIsNotNull();
+};